requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
billiard==4.1.0
celery==5.3.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
Django==4.2.2
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-formtools==2.4.1
django-phonenumber-field==6.4.0
django-timezone-field==5.1
django-two-factor-auth==1.15.3
flower==2.0.0
humanize==4.7.0
kombu==5.3.1
Pillow==10.0.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
pypng==0.20220715.0
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3
redis==4.6.0
six==1.16.0
sqlparse==0.4.4
tornado==6.3.2
typing_extensions==4.7.0
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
